# schemas/openapi/parties.yaml
openapi: 3.1.0
info:
  title: CREST parties (né registry, #50)
  version: "1"
  description: |
    ## Who may call what (#102)

    Every service verifies bearer tokens; this section is who each surface
    answers. Four kinds of path:

    - **Party-authorized** — acts in a named party's name; the caller must be
      that party or hold act-for-party where the action is scoped
      (X-CREST-On-Behalf-Of names the party). Marked `security: bearerAuth`.
    - **Authenticated** — any signed-in caller (custodian queues, operations
      lists). Also `bearerAuth`.
    - **Open by design** — verification, status list, instance, capability
      reads by unguessable id, and the bootstrap plane; each carries a comment
      saying why.
    - **`/internal/*`** — the other services' doors, unauthenticated by the
      service-identity ruling (§16, 2026-08-25): service traffic lives on
      internal routes and fencing them to the service network is the
      deployment's job.
    Who a party is, what they agreed to, what they are authorised to do, and — the
    part the rest of the system leans on — whether an identifier from a source
    system resolves to exactly one of them.

    Request and response bodies that are CREST primitives are referenced by the
    `$id` of their JSON Schema under `schemas/`, never restated here. A second copy
    of a primitive is the thing `schemas/README.md` exists to prevent: two
    hand-maintained definitions agree right up until they don't.
servers:
...
paths:
...
  /v1/resolve:
    get:
      operationId: resolve
      summary: Which party does this identifier belong to?
      description: >
        The endpoint the evidence pipeline calls for every row, and the one whose
        third answer is the design. 200 means exactly one candidate. 404 means
        nothing matched and the row belongs in the unclear queue. 409 means more
        than one party carries this identifier: a hold is recorded and no merge
        happens (W4). The 409 is a designed outcome, not a failure — picking the
        best candidate would be a merge, and `merges_without_confirmation = 0` is a
        monitored metric.
      parameters:
        - { name: kind, in: query, schema: { type: string }, description: Which key is being matched (roster id, hashed national id, phone). }
        - { name: value, in: query, required: true, schema: { type: string }, description: The key value. Required. }
        - { name: contextId, in: query, schema: { type: string }, description: Narrows a roster id to the context it was issued in. }
      responses:
        "200":
          description: Exactly one candidate, with which key matched and at what confidence.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Match" }
        "400":
          description: No `value` supplied.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "404":
          description: Nothing matched. The caller sends the row to the unclear queue rather than guessing.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "409":
          description: >
            More than one candidate. A hold has been written and is returned; it
            names the candidates so a human can decide. Nothing was merged (W4).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Hold" }
        "500": { $ref: "#/components/responses/Internal" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        An OIDC access token from the deployment's identity provider (#89, §4.1).
        In a deployment that is eSignet; locally it is tools/mocks/oidc, which
        mints real ES256 tokens against a real JWKS so the verification path
        under test is the production one.


        The provider's `sub` is never used directly. CREST re-salts it under a
        per-deployment key before anything stores or compares it, because #52
        and #63 are open against eSignet's pairwise subject — it partitions by
        relying party rather than by client, and derives without a
        per-deployment salt. Two deployments reading the same eSignet therefore
        cannot correlate their workers by subject.
...
  schemas:
...
    Error:
      type: object
      description: >
        The error body every CREST service returns (pkg/httpx). The message is
        written for whoever reads it, including eventually a worker being told why
        something is held.
      required: [code, message]
      properties:
        code: { type: string, description: A machine-actionable code, e.g. no_match, invalid_query, schema_violation. }
        message: { type: string }
        problems:
          type: array
          items: { type: string }
          description: Per-field detail from schema validation, when there is any.
    Match:
      type: object
      description: One resolved candidate, with the evidence for the match.
      properties:
        partyId: { type: string }
        matchedOn: { type: string, description: Which key matched. }
        confidence: { type: number }
    Hold:
      type: object
      description: >
        A recorded decision not to merge. Written whenever resolve finds more than
        one candidate; the row is what someone later acts on (W4).


        The colliding value itself is deliberately absent — §16 asks that
        duplicate holds show existence rather than content, and the reason is
        concrete: for a contact-route hold the value is a phone number, and a
        queue anybody can list would otherwise hand out the phone number of
        every worker two records disagree about. `keyKind` says which kind of
        identifier collided, which is what a custodian needs to decide it.
      properties:
        id: { type: string }
        keyKind: { type: string, enum: [national-id-hash, contact-route, roster-id] }
        candidates: { type: array, items: { type: string } }
        reason: { type: string }
        createdAt: { type: string, format: date-time }
  responses:
...
    Internal:
      description: The service logged the cause and did not leak it.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...